requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-google-genai
faiss-cpu
youtube-transcript-api
python-dotenv
langchain-text-splitters
requests